test: integration test for local auth
