+++ frontend/appflowy_flutter/integration_test/desktop/settings/settings_runner.dart
@@ -1,9 +1,11 @@
 import 'package:integration_test/integration_test.dart';
 
 import 'notifications_settings_test.dart' as notifications_settings_test;
+import 'settings_billing_test.dart' as settings_billing_test;
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
   notifications_settings_test.main();
+  settings_billing_test.main();
 }
